MediaManager is finished

+++ src/main/java/com/workfront/internship/business/MediaManagerImpl.java
@@ -22,35 +22,52 @@ public class MediaManagerImpl implements MediaManager {
         mediaDao = new MediaDaoImpl(dataSource);
     }
     public List<Media> getMediaByProductID(int id){
+        if(id <= 0)
+            throw new RuntimeException("invalid id");
 
         List<Media> medias = mediaDao.getMediaByProductID(id);
         return medias;
     }
     public Media getMediaByMediaID(int id){
+        if(id <= 0)
+            throw new RuntimeException("invalid id");
 
         Media media = mediaDao.getMediaByMediaID(id);
         return media;
     }
     public int insertMedia(Media media){
+        if(!validateMedia(media))
+            throw new RuntimeException("invalid id");
 
         int index = mediaDao.insertMedia(media);
         return index;
 
     }
     public void updateMedia(Media media){
+        if(!validateMedia(media))
+            throw new RuntimeException("invalid id");
 
         mediaDao.updateMedia(media);
 
     }
     public void deleteMediaByID(int id){
+        if(id <= 0)
+            throw new RuntimeException("invalid id");
 
         mediaDao.deleteMediaByID(id);
 
     }
 
     public void deleteMediaByProductID(int id){
+        if(id <= 0)
+            throw new RuntimeException("invalid id");
 
         mediaDao.deleteMediaByProductID(id);
+    }
+    private boolean validateMedia(Media media){
+        if(media != null && media.getProductID() >=0 && media.getMediaPath() != null)
+            return true;
+        return false;
     }
 
 }
